modules/location consumer service/requirements.txt
kafka-python
SQLAlchemy==1.3.19